+++ src/utils/create-snapshots.ts
@@ -2,7 +2,8 @@ import * as ts from 'typescript';
 import {Snapshots} from '../definitions';
 import {traverse_node} from './traverse-node';
 
-export const create_snapshots = (program: ts.Program, source_filename: string, lines: number[]) => {
+export const create_snapshots = (
+    program: ts.Program, source_filename: string, lines: number[], flag: ts.TypeFormatFlags) => {
   const source_file = program.getSourceFile(source_filename);
   const snapshots: Snapshots = {};
 
@@ -32,7 +33,7 @@ export const create_snapshots = (program: ts.Program, source_filename: string, l
     if (line_index !== -1) {
       const target_node = node.getChildAt(0);
       const type = checker.getTypeAtLocation(target_node);
-      snapshots[trigger_line] = checker.typeToString(type, node, ts.TypeFormatFlags.NoTruncation);
+      snapshots[trigger_line] = checker.typeToString(type, node, flag);
       rest_lines.splice(line_index, 1);
     }
   });